warning: use of a signed integer operand with a binary bitwise operator [hicpp-signed-bitwise] - header - 2

+++ sdk/inc/azure/core/internal/az_retry_internal.h
@@ -14,7 +14,7 @@ AZ_NODISCARD AZ_INLINE int32_t _az_retry_calc_delay(
     int32_t max_retry_delay_msec)
 {
   int32_t const exponential_retry_after
-      = retry_delay_msec * (attempt <= 30 ? (1 << attempt) : INT32_MAX); // scale exponentially
+      = retry_delay_msec * (attempt <= 30 ? (1U << (uint32_t)attempt) : INT32_MAX); // scale exponentially
 
   return exponential_retry_after > max_retry_delay_msec ? max_retry_delay_msec
                                                         : exponential_retry_after;
